+++ code/Lib/variableValues/dialogs/BatchValidator.py
@@ -5,33 +5,37 @@ reload(variableValues.validation)
 import os
 import AppKit
 from vanilla import *
+from mojo.UI import OutputWindow
 from mojo.roboFont import OpenFont, OpenWindow
 from variableValues.validation import validateFonts
 
 
 class BatchValidator:
-    
+
+    # TO-DO: add validation of font-level data
+
     title       = 'BatchValidator'
     padding     = 10
     lineHeight  = 22
     buttonWidth = 100
     verbose     = True
+
     _sources    = {}
     _targets    = {}
 
     def __init__(self):
         self.w = FloatingWindow(
-                (600, self.lineHeight*13 + self.padding*5),
-                title=self.title)
+                (600, 400), title=self.title,
+                minSize=(200, 300))
 
-        x = y = p = self.padding
-        self.w.sourcesLabel = TextBox(
+        _groupSources = Group((0, 0, -0, -0))
+        x = y = p = 0
+        _groupSources.sourcesLabel = TextBox(
                 (x, y, -p, self.lineHeight),
                 'source fonts')
-
         y += self.lineHeight + p/2
-        self.w.sources = List(
-                (x, y, -p, self.lineHeight*3),
+        _groupSources.sources = List(
+                (x, y, -p, -p),
                 [],
                 allowsMultipleSelection=False,
                 allowsEmptySelection=False,
@@ -42,14 +46,14 @@ class BatchValidator:
                     callback=self.dropSourcesCallback),
                 )
 
-        y += self.lineHeight*3 + p
-        self.w.targetsLabel = TextBox(
+        _groupTargets = Group((0, 0, -0, -0))
+        x = y = p = 0
+        _groupTargets.targetsLabel = TextBox(
                 (x, y, -p, self.lineHeight),
                 'target fonts')
-
         y += self.lineHeight + p/2
-        self.w.targets = List(
-                (x, y, -p, self.lineHeight*7),
+        _groupTargets.targets = List(
+                (x, y, -p, -p),
                 [],
                 allowsMultipleSelection=True,
                 allowsEmptySelection=False,
@@ -60,7 +64,25 @@ class BatchValidator:
                     callback=self.dropTargetsCallback),
                 )
 
-        y += self.lineHeight*7 + p
+        self.w._splitDescriptors = [
+            dict(view=_groupSources, identifier="sources"),
+            dict(view=_groupTargets, identifier="targets"),
+        ]
+        x = y = p = self.padding
+        self.w.splitView = SplitView(
+                (x, y, -p, -(self.lineHeight + p*2)),
+                self.w._splitDescriptors,
+                isVertical=False,
+                dividerStyle='thick',
+            )
+
+        y = -(self.lineHeight + p)
+
+        # self.w.spinner = ProgressSpinner(
+        #         (x, y+4, 16, 16),
+        #         sizeStyle='small',
+        #         displayWhenStopped=False)
+
         self.w.validateButton = Button(
                 (-(self.buttonWidth+p), y, self.buttonWidth, self.lineHeight),
                 'validate',
@@ -75,11 +97,15 @@ class BatchValidator:
     # -------------
 
     @property
+    def sourcesList(self):
+        return self.w._splitDescriptors[0]['view'].sources
+
+    @property
     def sourceFont(self):
-        sourceSelection = self.w.sources.getSelection()
+        sourceSelection = self.sourcesList.getSelection()
         if not sourceSelection:
             return 
-        sourceItems = self.w.sources.get()
+        sourceItems = self.sourcesList.get()
         return sourceItems[sourceSelection[0]]
 
     @property
@@ -87,9 +113,13 @@ class BatchValidator:
         return self._sources[self.sourceFont]
 
     @property
+    def targetsList(self):
+        return self.w._splitDescriptors[1]['view'].targets
+
+    @property
     def targetFonts(self):
-        targetSelection = self.w.targets.getSelection()
-        targetItems = self.w.targets.get()
+        targetSelection = self.targetsList.getSelection()
+        targetItems = self.targetsList.get()
         targetFonts = []
         for i, targetFont in enumerate(targetItems):
             if i not in targetSelection:
@@ -122,8 +152,8 @@ class BatchValidator:
             for source in sources:
                 sourceName = os.path.splitext(os.path.split(source)[-1])[0]
                 self._sources[sourceName] = source
-                self.w.sources.append(sourceName)
-                self.w.sources.setSelection([0])
+                self.sourcesList.append(sourceName)
+                self.sourcesList.setSelection([0])
 
         return True
 
@@ -142,8 +172,8 @@ class BatchValidator:
             for target in targets:
                 targetName = os.path.splitext(os.path.split(target)[-1])[0]
                 self._targets[targetName] = target
-                self.w.targets.append(targetName)
-                self.w.targets.setSelection([0])
+                self.targetsList.append(targetName)
+                self.targetsList.setSelection([0])
 
         return True
 
@@ -177,126 +207,19 @@ class BatchValidator:
         targetFonts = [OpenFont(targetFontPath, showInterface=False) for targetFontPath in self.targetFontPaths]
         result = validateFonts(targetFonts, sourceFont)
 
+        # write output to the console
+        O = OutputWindow()
+        O.clear()
+        O.write(result)
+        O.show()
+
         # done
-        print(result)
         print('...done.\n')
 
 
 # ----
 # test
 # ----
-
-if __name__ == '__main__':
-
-    OpenWindow(BatchValidator)
-
-
-
-
-
-
-# import os
-# import AppKit
-# from vanilla import *
-# from mojo.roboFont import CurrentFont, OpenWindow
-
-
-# class BatchValidator:
-
-#     title  = 'BarchValidator'
-#     checks = ['width', 'points', 'components', 'anchors', 'unicodes']
-
-#     padding     = 10
-#     textHeight  = 20
-#     buttonWidth = 100
-
-#     def __init__(self):
-#         self.w = FloatingWindow(
-#                 (400, 600), self.title,
-#                 minSize=(200, 300))
-
-#         x = y = p = self.padding
-#         self.w.sourcesLabel = TextBox(
-#                 (x, y, -p, self.textHeight),
-#                 'source fonts')
-
-#         y += self.textHeight + p/2
-#         self.w.sources = List(
-#                 (x, y, -p, self.textHeight*3),
-#                 [],
-#                 allowsMultipleSelection=False,
-#                 allowsEmptySelection=False,
-#                 enableDelete=True,
-#                 otherApplicationDropSettings=dict(
-#                     type=AppKit.NSFilenamesPboardType,
-#                     operation=AppKit.NSDragOperationCopy,
-#                     callback=self.dropSourcesCallback),
-#             )
-
-#         y += self.textHeight*3 + p
-#         self.w.targetsLabel = TextBox(
-#                 (x, y, -p, self.textHeight),
-#                 'target fonts'
-#             )
-
-#         y += self.textHeight + p/2
-#         self.w.targets = List(
-#                 (x, y, -p, self.textHeight*3),
-#                 [],
-#                 allowsMultipleSelection=True,
-#                 allowsEmptySelection=False,
-#             )
-
-#         y += self.textHeight*3 + p
-#         # self.w._splitDescriptors = [
-#         #     dict(view=_groupSources, identifier="sources"),
-#         #     dict(view=_groupTargets, identifier="targets"),
-#         # ]
-#         # self.w.splitView = SplitView(
-#         #         (x, y, -p, -self.textHeight-p*2),
-#         #         self.w._splitDescriptors,
-#         #         isVertical=False,
-#         #         dividerStyle='thick',
-#         #     )
-
-#         # y = -(self.textHeight + p)
-#         self.w.validateButton = Button(
-#                 (-(self.buttonWidth+p), y, self.buttonWidth, self.textHeight),
-#                 'validate',
-#                 # callback=self.batchValidateFontsCallback,
-#                 sizeStyle='small',
-#             )
-
-#         self.w.getNSWindow().setTitlebarAppearsTransparent_(True)
-#         self.w.open()
-
-#     # # -------------
-#     # # dynamic attrs
-#     # # -------------
-
-#     # @property
-#     # def sourceFont(self):
-#     #     sourcesList = self.w._splitDescriptors[0]['view']#.list
-#     #     print(sourcesList)
-#     #     sourceSelection = sourcesList.getSelection()
-#     #     if not sourceSelection:
-#     #         return 
-#     #     sourcesItems = sourcesList.get()
-#     #     return sourcesItems[sourceSelection[0]]
-
-#     # @property
-#     # def targetFonts(self):
-#     #     targetList = self.w._splitDescriptors[1]['view'].list
-#     #     targetSelection = targetList.getSelection()
-#     #     targetItems = targetList.get()
-#     #     targetFonts = []
-#     #     for i, targetFont in enumerate(targetItems):
-#     #         if i not in targetSelection:
-#     #             continue
-#     #         targetFonts.append(targetFont)
-#     #     return targetFonts
-
-
 
 if __name__ == '__main__':