+++ code/Lib/variableValues/dialogs/GlyphValidator.py
@@ -1,245 +1,199 @@
-import os, sys, glob, csv
-from AppKit import NSFilenamesPboardType, NSDragOperationCopy
-from vanilla import *
-from mojo.UI import AccordionView
-from mojo.roboFont import *
+from vanilla import FloatingWindow, Button, TextBox, CheckBox
+from mojo import drawingTools as ctx
+from mojo.UI import UpdateCurrentGlyphView, GetFile
+from mojo.events import addObserver, removeObserver
+from mojo.roboFont import OpenWindow, CurrentFont
+from variableValues.validation import *
 
-'''
-A tool to check individual sources for internal consistency.
-
-'''
 
 class GlyphValidator:
-    
-    title        = 'GlyphValidator'
-    key          = 'com.hipertipo.glyphValidator'
-    width        = 123*5
-    height       = 640
-    padding      = 10
-    lineHeight   = 22
-    verbose      = True
-    buttonWidth  = 100
 
-    _colGlyphs   = 120
-    _colFontName = 180
-    _colValue    = 60
+    title = "validate"
+    padding = 10
+    lineHeight = 20
+    verbose = False
 
-    _tabsTitles  = ['definitions', 'validation']
-    _memesTitles = ['glyph name', 'width', 'left', 'right', 'body', 'counter'] 
+    defaultFont = None
 
-    _memesFolder = '/Users/sergiogonzalez/Desktop/hipertipo/tools/VariableValues/data/recipes' # '/hipertipo/tools/VarTools/_data/glyphMemes/'
-    # assert os.path.exists(_memesFolder)
-
-    _fonts = {}
+    checks = [
+        'width',
+        'points',
+        'components',
+        'anchors',
+        'unicodes',
+    ]
+    checkResults = {}
 
     def __init__(self):
-        self.w = FloatingWindow(
-                (self.width, self.height), title=self.title,
-                minSize=(self.width*0.9, 360))
+        h = self.lineHeight*(4+len(self.checks)) + self.padding*4
+        self.w = FloatingWindow((123, h), self.title)
 
         x = y = p = self.padding
-        self.w.tabs = Tabs((x, y, -p, -p), self._tabsTitles)
-
-        # glyph memes
-
-        tab1 = self.w.tabs[0]
-
-        x, y = p, p/2
-        tab1.memesFolderLabel = TextBox(
+        self.w.getDefaultButton = Button(
                 (x, y, -p, self.lineHeight),
-                'source folder')
-
-        y += self.lineHeight + p/2
-        tab1.memesFolder = EditText(
-                (x, y, -(self.buttonWidth+p*2), self.lineHeight),
-                self._memesFolder)
-        
-        tab1.memesGetFolder = Button(
-                (-(self.buttonWidth+p), y, self.buttonWidth, self.lineHeight),
-                'get folder')
-        
-        y2 = y + self.lineHeight + p
-        tab1.memesLabel = TextBox(
-                (x, y2, -p, self.lineHeight),
-                'glyph memes')
-
-        y2 += self.lineHeight + p/2
-        tab1.glyphs = List(
-                (x, y2, self._colGlyphs, -(self.lineHeight + p*2)),
-                [],
-                allowsMultipleSelection=False,
-                allowsEmptySelection=False,
-                selectionCallback=self.selectMemeCallback,
+                'get default...',
+                callback=self.getDefaultCallback,
+                sizeStyle='small',
             )
 
-        x2 = x + self._colGlyphs + p
-        tab1.memes = List(
-                (x2, y2, -p, -(self.lineHeight + p*2)),
-                [],
-                allowsMultipleSelection=False,
-                allowsEmptySelection=False,
-                columnDescriptions=[{"title": t} for t in self._memesTitles],
-            )
+        y += self.lineHeight + p
 
-        y = -(self.lineHeight + p)
-        tab1.updateValues = Button(
-                (x, y, self.buttonWidth, self.lineHeight),
-                'update',
-                # callback=self.updateMeasurementsCallback,
-            )
-
-        # validation
-
-        tab2 = self.w.tabs[1]
-
-        x, y = p, p/2
-        tab2.fontsLabel = TextBox(
+        self.w.checksLabel = TextBox(
                 (x, y, -p, self.lineHeight),
-                'fonts')
+                "glyph tests",
+                sizeStyle='small')
 
-        y += self.lineHeight + p/2
-        tab2.fonts = List(
-                (x, y, -p, self.lineHeight*5),
-                [],
-                allowsMultipleSelection=False,
-                allowsEmptySelection=False,
-                # enableDelete=True,
-                # editCallback=self.selectDesignspaceCallback,
-                # selectionCallback=self.selectDesignspaceCallback,
-                otherApplicationDropSettings=dict(
-                    type=NSFilenamesPboardType,
-                    operation=NSDragOperationCopy,
-                    callback=self.dropFontsCallback),
-            )
-
-        y += self.lineHeight*5 + p
-        tab2.memesLabel = TextBox(
+        y += self.lineHeight # + p
+        for check in self.checks:
+            checkbox = CheckBox(
                 (x, y, -p, self.lineHeight),
-                'validation results')
+                check,
+                value=True,
+                callback=self.updateFontViewCallback,
+                sizeStyle='small')
+            setattr(self.w, check, checkbox)
+            y += self.lineHeight
 
-        y2 = y + self.lineHeight + p/2
-        tab2.glyphs = List(
-                (x, y2, self._colGlyphs, -(self.lineHeight + p*2)),
-                [],
-                allowsMultipleSelection=False,
-                allowsEmptySelection=False,
-            )
+        y += p
+        self.w.markCells = CheckBox(
+                (x, y, -p, self.lineHeight),
+                "font marks",
+                value=True,
+                callback=self.updateFontViewCallback,
+                sizeStyle='small')
 
-        x2 = x + self._colGlyphs + p
-        tab2.memes = List(
-                (x2, y2, -p, -(self.lineHeight + p*2)),
-                [],
-                allowsMultipleSelection=False,
-                allowsEmptySelection=False,
-                columnDescriptions=[{"title": t} for t in self._memesTitles],
-                )
+        y += self.lineHeight
+        self.w.markGlyphs = CheckBox(
+                (x, y, -p, self.lineHeight),
+                "glyph marks",
+                value=True,
+                callback=self.updateGlyphViewCallback,
+                sizeStyle='small')
 
-        y = -(self.lineHeight + p)
-        tab2.updateValues = Button(
-                (x, y, self.buttonWidth, self.lineHeight),
-                'update',
-                # callback=self.updateMeasurementsCallback,
-            )
 
-        x += self.buttonWidth + p
-        tab2.visualizeValues = Button(
-                (x, y, self.buttonWidth, self.lineHeight),
-                'visualize',
-                # callback=self.visualizeMeasurementsCallback,
-            )
-
-        x += self.buttonWidth + p
-        tab2.exportValues = Button(
-                (x, y, self.buttonWidth, self.lineHeight),
-                'export',
-                # callback=self.visualizeFontinfoCallback,
-            )
-
-        self.loadMemes()
+        self.w.bind("close", self.closeCallback)
+        self.w.getNSWindow().setTitlebarAppearsTransparent_(True)
         self.w.open()
 
-    # -------------
-    # dynamic attrs
-    # -------------
+        addObserver(self, "updateGlyphChecks", "currentGlyphChanged")
+        addObserver(self, "drawLabelsCell",    "glyphCellDrawBackground")
+        addObserver(self, "drawLabelsGlyph",   "drawBackground")
 
-    @property
-    def selectedGlyph(self):
-        tab = self.w.tabs[0]
-        selection = tab.glyphs.getSelection()
-        glyphs = tab.glyphs.get()
-        selectedGlyphs = [a for i, a in enumerate(glyphs) if i in selection]
-        if not len(selectedGlyphs):
-            return
-        return selectedGlyphs[0]
-
-    # ---------
-    # callbacks
-    # ---------
-
-    def selectMemeCallback(self, sender):
-        csvPath = os.path.join(self._memesFolder, f'{self.selectedGlyph}.csv')
-        tab = self.w.tabs[0]
-
-        # reset list
-        memesPosSize = tab.memes.getPosSize()
-        del tab.memes
-
-        descriptions  = [{"title": self._memesTitles[0], 'width': self._colFontName*1.5, 'minWidth': self._colFontName}]
-        descriptions += [{"title": t, 'width': self._colValue} for t in self._memesTitles[1:]]
-        listItems = []
-
-        if os.path.exists(csvPath):
-            with open(csvPath, mode ='r') as f:
-                csvFile = csv.reader(f)
-                for lines in csvFile:
-                    item = {}
-                    for i, t in enumerate(self._memesTitles):
-                        item[t] = lines[i]
-                    listItems.append(item)
-
-        tab.memes = List(
-            memesPosSize,
-            listItems,
-            columnDescriptions=descriptions)
-
-    def dropFontsCallback(self, sender, dropInfo):
-        isProposal = dropInfo["isProposal"]
-        existingPaths = sender.get()
-
-        paths = dropInfo["data"]
-        paths = [path for path in paths if path not in existingPaths]
-        paths = [path for path in paths if os.path.splitext(path)[-1].lower() == '.ufo']
-
-        if not paths:
-            return False
-
-        if not isProposal:
-            tab = self.w.tabs[1]
-            for path in paths:
-                fileName = os.path.splitext(os.path.split(path)[-1])[0]
-                self._fonts[fileName] = path
-                tab.fonts.append(fileName)
-                tab.fonts.setSelection([0])
-
-        return True
-
-    # -------
     # methods
-    # -------
 
-    def loadMemes(self):
-        tab = self.w.tabs[0]
-        csvFolder = tab.memesFolder.get()
-        csvFiles = glob.glob(f'{csvFolder}/*.csv')
-        glyphNames = sorted([os.path.splitext(os.path.split(f)[-1])[0] for f in csvFiles])
-        tab.glyphs.set(glyphNames)
-        # load glyph names in validation tab
-        tab2 = self.w.tabs[1]
-        tab2.glyphs.set(glyphNames)
+    def checkGlyph(self, g1):
+        if g1 is None:
+            return
 
+        if self.defaultFont is None:
+            if self.verbose:
+                print('no default source selected.')
+            return
+
+        if g1.name not in self.defaultFont:
+            if self.verbose:
+                print(f"glyph '{g1.name}' not in default source.")
+            return
+
+        g2 = self.defaultFont[g1.name]
+
+        self.checkResults = {
+            'width'      : validateWidth(g1, g2),
+            'points'     : validateContours(g1, g2),
+            'components' : validateComponents(g1, g2),
+            'anchors'    : validateAnchors(g1, g2),
+            'unicodes'   : validateUnicodes(g1, g2),
+        }
+
+    # callbacks
+
+    def getDefaultCallback(self, sender):
+        defaultPath = GetFile(message='Get default source…', title=self.title)
+        self.defaultFont = OpenFont(defaultPath, showInterface=False)
+        self.updateGlyphViewCallback(sender)
+        self.updateFontViewCallback(sender)
+
+    def updateGlyphViewCallback(self, sender):
+        UpdateCurrentGlyphView()
+
+    def updateFontViewCallback(self, sender):
+        font = CurrentFont()
+        if font is None:
+            return
+        for g in font:
+            g.changed()
+
+    def closeCallback(self, sender):
+        font = CurrentFont()
+        if font is None:
+            return
+        removeObserver(self, "glyphCellDrawBackground")
+        removeObserver(self, "drawBackground")
+        removeObserver(self, "currentGlyphChanged")
+        self.updateGlyphViewCallback(sender)
+        self.updateFontViewCallback(sender)
+
+    # observers
+
+    def updateGlyphChecks(self, notification):
+        glyph = notification['glyph']
+        self.checkGlyph(glyph)
+
+    def drawLabelsCell(self, notification):
+        if not self.w.markCells.get():
+            return
+
+        glyph = notification['glyph']
+        
+        self.checkGlyph(glyph)
+
+        if not len(self.checks):
+            return
+
+        ctx.save()
+        ctx.font('Menlo-Bold')
+        ctx.fontSize(10)
+        ctx.translate(3, 3)
+        for check in self.checkResults.keys():
+            checkbox = getatttr(self.w, check)
+            print(check, checkbox)
+            print(checkbox.get())
+            if checkbox.get() is not True:
+                continue
+            if self.checkResults[check]:
+                ctx.fill(0, 1, 0)
+            else:
+                ctx.fill(1, 0, 0)
+            ctx.text(check[0].upper(), (0, -3))
+            w, h = ctx.textSize(check[0].upper())
+            ctx.translate(w+2, 0)
+        ctx.restore()
+
+    def drawLabelsGlyph(self, notification):
+        if not self.w.markGlyphs.get():
+            return
+
+        glyph = notification['glyph']
+        scale = notification['scale']
+
+        font = glyph.font
+
+        self.checkGlyph(glyph)
+
+        ctx.save()
+        ctx.fontSize(12*scale)
+        for check in self.checkResults.keys():
+            if self.checkResults[check] is True:
+                ctx.fill(0, 1, 0.2)
+            else:
+                ctx.fill(1, 0, 0)
+            label = check[0].upper()
+            ctx.text(label, (4, 4))
+            w, h = ctx.textSize(label)
+            ctx.translate(w + 2*scale, 0)
+        ctx.restore()
 
 
 if __name__ == '__main__':
 
     OpenWindow(GlyphValidator)
-